Fix for torch.jit.script patching applied to classes (#1626)
Generally `torch.jit.script` is applied to methods, but in some cases it may be applied to classes, for example in `kornia` package. In such case NNCF patching of `torch.jit.script` is incorrect and leads to an error during import of the scripted class. 

The problem arises because patching inserts and additional item to call stack and `torch.jit.script` relies on it when retrieving `ResourceResolutionCallback` object representing the scopes for the trace methods of the class. All this results in the need for providing a custom `ResourceResolutionCallback` during patching.

Also, `ORIGINAL_OPERATORS` list was increasing in size after each unpatch-patch iteration resulting in duplication up to 10x. This is because it wasn't cleared before yet another patching operation. This issue was fixed in this PR too.

### Related tickets

81980

### Tests

Added a test for the case when a class is being scripted.